combines auth api calls in one file
